backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Nirmal Joji",email = "nirmalsharon.joji@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "langchain-core (>=0.3.41,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "assistant-stream (>=0.0.14,<0.0.15)",
    "motor (>=3.7.0,<4.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langchain (>=0.3.20,<0.4.0)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "chardet (>=5.2.0,<6.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.1,<3.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"